+++ app/Http/Controllers/AppointmentController.php
@@ -12,7 +12,7 @@ class AppointmentController extends Controller
 {
   private $appointmentService;
   
-   /**
+  /**
     * Service For Appointment 
     *
     * App\Service\Appointment Service
@@ -22,48 +22,27 @@ class AppointmentController extends Controller
     $this->appointmentService=$appointmentService;
   }
 
-  /**
-    * Show all project related appointments 
-    *
-    * App\Service\Appointment Service
-    */
   public function index(Project $project,Request $request)
   {
     return  $project->appointments()->with('users')->get();
   }
 
-  /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     */
   public function store(Project $project,AppointmentRequest $request)
   {
     $appointment=$project->appointments()->create($request->validated());
 
-    $this->appointmentService->performAppointmentRelatedTasks($project,$appointment);    
+    $this->appointmentService->performAppointmentRelatedTasks($project,$appointment);  
   }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  int  $project, int $appointment
-     * @return  \Illuminate\Http\Request  $request
-     *
-     */
     public function update(Project $project,Appointment $appointment,AppointmentRequest 
       $request)
     {
       $appointment->update($request->validated());
 
-      $this->appointmentService->performRelatedOperation($project,$request,$appointment);
+      $this->appointmentService->
+      performRelatedOperation($project,$request,$appointment);
     }
 
-    /**
-     * Delete the specified resource from database.
-     *
-     * @param  int  $project, int $appointment
-     */
      public function destroy(Project $project,Appointment $appointment)
      {
        $appointment->delete();
