+++ app/Http/Controllers/HomeController.php
@@ -17,23 +17,11 @@ class HomeController extends Controller
         $this->middleware('auth');
     }
 
-    /**
-     * Show the application dashboard.
-     *
-     * @return \Illuminate\Contracts\Support\Renderable
-     */
     public function index()
     {
       return view('home');
     }
 
-
-    /**
-     * Return All Users.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function users()
     {
         return User::latest()->get();
