+++ app/Http/Views/Composers/ProjectComposer.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Http\View\Composers;
+
+use App\Repositories\UserRepository;
+use Illuminate\View\View;
+
+class ProjectComposer
+{
+    /**
+     * The user repository implementation.
+     *
+     * @var \App\Repositories\UserRepository
+     */
+    protected $project;
+
+    /**
+     * Create a new profile composer.
+     *
+     * @param  \App\Repositories\UserRepository  $users
+     * @return void
+     */
+    public function __construct(UserRepository $users)
+    {
+        // Dependencies are automatically resolved by the service container...
+        $this->project = $project;
+    }
+
+    /**
+     * Bind data to the view.
+     *
+     * @param  \Illuminate\View\View  $view
+     * @return void
+     */
+    public function compose(View $view)
+    {
+        $view->with('lastSeen', $this->user->lastseen);
+    }
+}
