+++ app/Http/Views/Composers/ViewServiceProvider.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Providers;
+
+use App\Http\View\Composers\ProfileComposer;
+use Illuminate\Support\Facades\View;
+use Illuminate\Support\ServiceProvider;
+
+class ViewServiceProvider extends ServiceProvider
+{
+    /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        //
+    }
+
+    /**
+     * Bootstrap any application services.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        // Using class based composers...
+        View::composer('profile', ProfileComposer::class);
+
+        // Using closure based composers...
+        View::composer('dashboard', function ($view) {
+            //
+        });
+    }
+}
