+++ app/RecordActivity.php
@@ -69,9 +69,9 @@ trait RecordActivity
     public function activity()
     {
         if (get_class($this) === Project::class) {
-            return $this->hasMany(Activity::class)->latest();
+            return $this->hasMany(Activity::class)->with('user')->latest();
         }
-        return $this->morphMany(Activity::class, 'subject')->latest();
+        return $this->morphMany(Activity::class, 'subject')->with('user')->latest();
     }
     /**
      * Fetch the changes to the model.
