+++ app/Services/ConversationService.php
@@ -7,12 +7,7 @@ use App\User;
 
 class ConversationService extends \App\Http\Controllers\Controller
 {
-    /**
-     *  Store file conversation.
-     *
-     * @param  int  $project
-     * @return \Illuminate\Http\Response
-     */
+
   public function storeFileConversation($project,$request)
   {
     $this->validate($request, [
@@ -24,12 +19,6 @@ class ConversationService extends \App\Http\Controllers\Controller
     return $this->createConversation($project,'file',$value);
   }
  
-    /**
-     *  Store static conversation.
-     *
-     * @param  int  $project
-     * @return \Illuminate\Http\Response
-     */
   public function storeStaticConversation($project,$request)
   {
     $this->validate($request, [
@@ -39,18 +28,11 @@ class ConversationService extends \App\Http\Controllers\Controller
    return $this->createConversation($project,'message',request('message'));
   }
 
-    /**
-     *  Create conversation.
-     *
-     * @param  int  $project,string $value
-     * @return \Illuminate\Http\Response
-     */
   public function createConversation($project,$name,$value)
   {
-    return $conversation = Conversation::create([
-      $name => $value,
-      'group_id' => $project->group->id,
-      'user_id' => auth()->id(),
+    return $project->group->conversations()->create([
+        $name => $value,
+        'user_id' => auth()->id(),
       ]);
   }
 }
