+++ composer.json
@@ -9,7 +9,9 @@
     "license": "MIT",
     "require": {
         "php": "^7.2.5",
+        "barryvdh/laravel-debugbar": "^3.5",
         "fideloper/proxy": "^4.2",
+        "friendsofphp/php-cs-fixer": "^2.18",
         "fruitcake/laravel-cors": "^1.0",
         "guzzlehttp/guzzle": "^6.3",
         "intervention/image": "^2.5",
@@ -28,6 +30,7 @@
         "twilio/sdk": "^6.11"
     },
     "require-dev": {
+        "beyondcode/laravel-query-detector": "^1.5",
         "facade/ignition": "^2.0",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
