+++ extra
@@ -54,3 +54,6 @@ https://crm-fresh.s3.ap-south-1.amazonaws.com/3_60...
        paypal_yearly:P-94D04691SW198613W4RP2ZQQ       
 
        active_random:P-7EX67166FE871724S2QJSFAY
+
+                   <!--@foreach($project->activity->take(5) as $activity)-->
+             <!--@endforeach-->
