+++ resources/views/project/activities/accept_member_project.blade.php
@@ -1,4 +1,7 @@
 <span class="activity-icon activity-icon_purple"><i class="fas fa-user"></i></span>
+
  <a href='/users{{strrchr($activity->detail, '/_/')}}/profile' target="_blank">
+
   {{strtok($activity->detail, '/_/')}}
+  
 </a> Accepted project member invitation
