+++ resources/views/project/activities/activities.blade.php
@@ -26,11 +26,22 @@
                  </div>
                  <div class="card-body activity-search">
                    <ul>
-                     <li><a href="/projects/{{$project->id}}/timeline_feeds" class="{{ Route::is('activities') && empty(Request::query()) ? 'activity-icon_secondary activity-font' : '' }}"><i class="fas fa-layer-group activity-icon_secondary mr-3"></i>All Activities</a></li>
-                     <li><a href="/projects/{{$project->id}}/timeline_feeds?mine={{auth()->user()->id}}" class="{{ request()->has('mine') ? 'activity-icon_purple activity-font' : '' }}"><i class="fas fa-user activity-icon_purple mr-3"></i> My Activities</a></li>
-                     <li><a href="/projects/{{$project->id}}/timeline_feeds?related=1" class="{{ request()->has('related') ? 'activity-icon_green activity-font' : '' }}"><i class="far fa-star activity-icon_green mr-3"></i> Project Activities</a></li>
-                     <li><a href="/projects/{{$project->id}}/timeline_feeds?task=1" class="{{ request()->has('task') ? 'activity-icon_primary activity-font' : '' }}"><i class="fas fa-tasks activity-icon_primary mr-3"></i> Task Activities</a></li>
-                     <li><a href="/projects/{{$project->id}}/timeline_feeds?appointment=1" class="{{ request()->has('appointment') ? 'activity-icon_appoint activity-font' : '' }}"><i class="far fa-calendar-check activity-icon_appoint mr-3"></i> Appointment Activities</a></li>
+                     <li>
+                        <a href="/projects/{{$project->id}}/timeline_feeds" class="{{ Route::is('activities') && empty(Request::query()) ? 'activity-icon_secondary activity-font' : '' }}"><i class="fas fa-layer-group activity-icon_secondary mr-3"></i>All Activities</a>
+                    </li>
+
+                     <li>
+                        <a href="/projects/{{$project->id}}/timeline_feeds?mine={{auth()->user()->id}}" class="{{ request()->has('mine') ? 'activity-icon_purple activity-font' : '' }}"><i class="fas fa-user activity-icon_purple mr-3"></i> My Activities</a>
+                    </li>
+                     <li>
+                        <a href="/projects/{{$project->id}}/timeline_feeds?related=1" class="{{ request()->has('related') ? 'activity-icon_green activity-font' : '' }}"><i class="far fa-star activity-icon_green mr-3"></i> Project Activities</a>
+                    </li>
+                     <li>
+                        <a href="/projects/{{$project->id}}/timeline_feeds?task=1" class="{{ request()->has('task') ? 'activity-icon_primary activity-font' : '' }}"><i class="fas fa-tasks activity-icon_primary mr-3"></i> Task Activities</a>
+                    </li>
+                     <li>
+                        <a href="/projects/{{$project->id}}/timeline_feeds?appointment=1" class="{{ request()->has('appointment') ? 'activity-icon_appoint activity-font' : '' }}"><i class="far fa-calendar-check activity-icon_appoint mr-3"></i> Appointment Activities</a>
+                    </li>
                    </ul>
                  </div>
                 </div>
@@ -40,7 +51,5 @@
             </div>
         </div>
     </div>
-
-
 
 @endsection
